chore: 풀이 일부 변경

+++ src/programmers/level2/PGS_멀쩡한사각형.java
@@ -18,8 +18,8 @@ public class PGS_멀쩡한사각형 {
         long answer = (long)w * h;
 
         for (long i = 1; i <= w; i++) {
-            long before = (long)Math.floor((i - 1) * (long)h/(double)w);
-            long after = (long)Math.ceil(i * (long)h/(double)w);
+            long before = (long)Math.floor((i - 1) * h/(double)w);
+            long after = (long)Math.ceil(i * h/(double)w);
 
             answer -= after - before;
         }
